refactor: replace head.tsx to metadata api and refactor layout
